refactor: split settings-dialog.tsx into modular components (Phase 3, Task 3)
Split components/settings-dialog.tsx from 454 lines into focused modules:
- settings/about-section.tsx (69 lines) - Version and app information
- settings/appearance-settings.tsx (91 lines) - Theme and display preferences
- settings/data-management.tsx (107 lines) - Import/export and storage stats
- settings/notification-settings.tsx (113 lines) - Notification configuration
- settings/settings-dialog.tsx (169 lines) - Main dialog wrapper with state
- settings-dialog.tsx (15 lines) - Re-export layer for backward compatibility

Benefits:
- All files now under 170 lines (well below 300-line limit)
- Single responsibility per section component
- Improved maintainability and testability
- Zero breaking changes (100% backward compatible)

Verification:
- TypeScript compilation: No new errors
- All existing imports work unchanged
- Each section independently testable

Related: Phase 3 of coding standards compliance (see next.md)
